Move AchievementCard component into flare directory
